requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file

transformers
langchain
sentence-transformers
faiss-cpu
google-auth
google-api-python-client
google-auth-oauthlib